FastApi/requirements.txt
alembic
fastapi
insightface
numpy
onnxruntime
opencv-python
pandas
passlib
psycopg2
python-dotenv
python-jose
python-multipart
scikit-learn
sqlalchemy
uvicorn
websockets
asyncpg
sqladmin
apscheduler